requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==4.4.2
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.9
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.2.1
kombu==4.6.11
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
prompt-toolkit==3.0.22
PyJWT==2.3.0
pyparsing==3.0.6
pyrsistent==0.18.0
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
uritemplate==4.1.1
urllib3==1.26.7
vine==1.3.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
